+++ tests/test_loader.py
@@ -2,10 +2,13 @@
 import json
 import pytest
 import requests_mock
+import tempfile
 
 from chaoslib.exceptions import InvalidSource
-from chaoslib.loader import load_experiment
+from chaoslib.loader import load_experiment, parse_experiment_from_file
 from chaoslib.types import Settings
+
+from fixtures import experiments
 
 
 def test_load_from_file(generic_experiment: str):
@@ -61,3 +64,23 @@ def test_load_from_http_with_auth(settings: Settings, generic_experiment: str):
             load_experiment('http://example.com/experiment.json', settings)
         except InvalidSource as x:
             pytest.fail(str(x))
+
+
+def test_yaml_safe_load_from_file():
+    with tempfile.NamedTemporaryFile(suffix=".yaml") as f:
+        f.write(experiments.UnsafeYamlExperiment.encode('utf-8'))
+        f.seek(0)
+
+        with pytest.raises(InvalidSource):
+            parse_experiment_from_file(f.name)
+
+
+def test_yaml_safe_load_from_http():
+    with requests_mock.mock() as m:
+        m.get(
+            'http://example.com/experiment.yaml', status_code=200,
+            headers={"Content-Type": "application/x-yaml"},
+            text=experiments.UnsafeYamlExperiment
+        )
+        with pytest.raises(InvalidSource):
+            load_experiment('http://example.com/experiment.yaml')